move segmeta read/write to sep file, move out allpqsids, avoid cmi re-reads (#1744)
* move all segmeta reading and writing to a new module

* read and write pqids and columns names into a separate file

* lint

* move backfillpqssegmeta to its own file

* move deleting of pqsdata to segmetarw

* misc fixes

* fix index delete scenario

* fix recursive locking

* remove lock for emptypqsmeta file updating

* fix pqmeta filename

* add init func for pqsmeta as an explicit one

* some debug logs

* dont decode pqsmeta if filesize is 0

* avoid re-reading cmi if it is already presnet for the blk/cname

* create an explicit func for sendErrorStatus

* increase num of backups to 50

* add some detailed log and fixed var reuse in goroutine

* remove unused tablename, don't re-read and re-populate pqid in allpersistentqueryresults

* simplify persistentqueryresults struct

* add pqids for a given segkey in one shot

* better error msg when reading emptypqs file

* Update pkg/segment/query/metadata/unrotatedmeta.go


* delete segment validity file last

* pr comments

* better error handling for writing emptypqsmeta file

* better log

---------

+++ pkg/retention/retention.go
@@ -21,6 +21,7 @@ import (
 	"math"
 	"path"
 	"sort"
+	"strings"
 	"time"
 
 	"github.com/dustin/go-humanize"
@@ -80,12 +81,7 @@ func DoRetentionBasedDeletion(ingestNodeDir string, retentionHours int, orgid ui
 	currTime := time.Now()
 	deleteBefore := GetRetentionTimeMs(retentionHours, currTime)
 
-	currentSegmeta := path.Join(ingestNodeDir, writer.SegmetaSuffix)
-	allSegMetas, err := writer.ReadSegmeta(currentSegmeta)
-	if err != nil {
-		log.Errorf("DoRetentionBasedDeletion: Failed to read segmeta, filePath=%v, err: %v", currentSegmeta, err)
-		return
-	}
+	allSegMetas := writer.ReadLocalSegmeta(false)
 
 	// Read metrics meta entries
 	currentMetricsMeta := path.Join(ingestNodeDir, mmeta.MetricsMetaSuffix)
@@ -158,7 +154,7 @@ func DoRetentionBasedDeletion(ingestNodeDir string, retentionHours int, orgid ui
 		len(allEntries), len(segmentsToDelete), len(metricSegmentsToDelete), oldest, orgid)
 
 	// Delete all segment data
-	DeleteSegmentData(currentSegmeta, segmentsToDelete, true)
+	DeleteSegmentData(segmentsToDelete, true)
 	DeleteMetricsSegmentData(currentMetricsMeta, metricSegmentsToDelete, true)
 	DeleteEmptyIndices(ingestNodeDir, orgid)
 }
@@ -170,13 +166,8 @@ func DeleteEmptyIndices(ingestNodeDir string, myid uint64) {
 		return
 	}
 
-	currentSegmeta := path.Join(ingestNodeDir, writer.SegmetaSuffix)
+	segMetaEntries := writer.ReadLocalSegmeta(false)
 
-	segMetaEntries, err := writer.ReadSegmeta(currentSegmeta)
-	if err != nil {
-		log.Errorf("deleteEmptyIndices: Error in reading segmeta file=%v, err: %v", currentSegmeta, err)
-		return
-	}
 	// Create a set of virtualTableName values from segMetaEntries
 	virtualTableNames := make(map[string]struct{})
 	for _, entry := range segMetaEntries {
@@ -233,12 +224,7 @@ func doVolumeBasedDeletion(ingestNodeDir string, allowedVolumeGB uint64, deletio
 		return
 	}
 
-	currentSegmeta := path.Join(ingestNodeDir, writer.SegmetaSuffix)
-	allSegMetas, err := writer.ReadSegmeta(currentSegmeta)
-	if err != nil {
-		log.Errorf("doVolumeBasedDeletion: Failed to read segmeta, filepath=%v, err: %v", currentSegmeta, err)
-		return
-	}
+	allSegMetas := writer.ReadLocalSegmeta(false)
 
 	currentMetricsMeta := path.Join(ingestNodeDir, mmeta.MetricsMetaSuffix)
 	allMetricMetas, err := mmeta.ReadMetricsMeta(currentMetricsMeta)
@@ -297,14 +283,14 @@ func doVolumeBasedDeletion(ingestNodeDir string, allowedVolumeGB uint64, deletio
 			}
 		}
 	}
-	DeleteSegmentData(currentSegmeta, segmentsToDelete, true)
+	DeleteSegmentData(segmentsToDelete, true)
 	DeleteMetricsSegmentData(currentMetricsMeta, metricSegmentsToDelete, true)
 }
 
 func getSystemVolumeBytes() (uint64, error) {
 	currentVolume := uint64(0)
 
-	allSegmetas := writer.ReadAllSegmetas()
+	allSegmetas := writer.ReadGlobalSegmetas()
 
 	allCnts := writer.GetVTableCountsForAll(0, allSegmetas)
 	writer.GetUnrotatedVTableCountsForAll(0, allCnts)
@@ -331,7 +317,7 @@ func getSystemVolumeBytes() (uint64, error) {
 	return currentVolume, nil
 }
 
-func DeleteSegmentData(segmetaFile string, segmentsToDelete map[string]*structs.SegMeta, updateBlob bool) {
+func DeleteSegmentData(segmentsToDelete map[string]*structs.SegMeta, updateBlob bool) {
 
 	if len(segmentsToDelete) == 0 {
 		return
@@ -344,8 +330,12 @@ func DeleteSegmentData(segmetaFile string, segmentsToDelete map[string]*structs.
 		// Delete segment files from s3
 		dirPath := segMetaEntry.SegmentKey
 		filesToDelete := fileutils.GetAllFilesInDirectory(path.Dir(dirPath) + "/")
-
+		svFileName := ""
 		for _, file := range filesToDelete {
+			if strings.Contains(file, utils.SegmentValidityFname) {
+				svFileName = file
+				continue
+			}
 			err := blob.DeleteBlob(file)
 			if err != nil {
 				log.Errorf("deleteSegmentData: Error in deleting segment file %v in blob, err: %v",
@@ -353,9 +343,17 @@ func DeleteSegmentData(segmetaFile string, segmentsToDelete map[string]*structs.
 				continue
 			}
 		}
+		if svFileName != "" {
+			err := blob.DeleteBlob(svFileName)
+			if err != nil {
+				log.Errorf("deleteSegmentData: Error deleting validity file %v in blob, err: %v",
+					svFileName, err)
+			}
+		}
+		log.Infof("DeleteSegmentData: deleted seg: %v", segMetaEntry.SegmentKey)
 	}
 
-	writer.RemoveSegments(segmetaFile, utils.MapToSet(segmentsToDelete))
+	writer.RemoveSegments(utils.MapToSet(segmentsToDelete))
 
 	// Upload the latest ingest nodes dir to s3 only if updateBlob is true
 	if !updateBlob {
